fix(coverage): move metrics TestProvider to runtime/command _test.go helper
CI kernel coverage gate failed: kernel/observability/metrics dropped to
36.1% because testprovider.go (non-_test.go, used only cross-package by
runtime/command) added uncovered statements to that production package.
TestProvider has exactly one consumer (lifecycle_test.go); relocate it as
an unexported package-command _test.go helper (coverage-neutral; no new
subpackage; no CI gate-exclusion edit). kernel/observability/metrics back
to 96.3%; gate PASS.
